Move grid package to cmp package
