+++ admin/tabs/services/ServiceModel.js
@@ -7,7 +7,7 @@
 import {XH, HoistModel} from 'hoist/core';
 import {ToastManager} from 'hoist/cmp';
 import {UrlStore} from 'hoist/data';
-import {GridModel} from 'hoist/grid';
+import {GridModel} from 'hoist/cmp/grid';
 import {baseCol} from 'hoist/columns/Core';
 
 @HoistModel()
