+++ admin/tabs/users/UserModel.js
@@ -7,7 +7,7 @@
 
 import {HoistModel} from 'hoist/core';
 import {UrlStore} from 'hoist/data';
-import {GridModel} from 'hoist/grid';
+import {GridModel} from 'hoist/cmp/grid';
 
 import {baseCol, boolCheckCol} from 'hoist/columns/Core';
 import {usernameCol} from '../../columns/Columns';
