+++ cmp/leftRightChooser/LeftRightChooser.js
@@ -7,7 +7,7 @@
 import {Component} from 'react';
 import {elemFactory, HoistComponent} from 'hoist/core';
 import {hframe, vframe} from 'hoist/layout';
-import {grid} from 'hoist/grid';
+import {grid} from 'hoist/cmp/grid';
 
 import {description} from './impl/Description';
 import {chooserToolbar} from './impl/ChooserToolbar';
