+++ rest/RestGrid.js
@@ -8,7 +8,7 @@
 import {Component} from 'react';
 import {defaults} from 'lodash';
 import {HoistComponent, elemFactory} from 'hoist/core';
-import {grid} from 'hoist/grid';
+import {grid} from 'hoist/cmp/grid';
 import {fragment} from 'hoist/layout';
 import {message, panel} from 'hoist/cmp';
